Refactor send-mail module
