+++ fractal_server/app/security/__init__.py
@@ -48,13 +48,11 @@ from fractal_server.app.models import OAuthAccount
 from fractal_server.app.models import UserGroup
 from fractal_server.app.models import UserOAuth
 from fractal_server.app.schemas.user import UserCreate
-from fractal_server.app.security.signup_email import (
-    send_fractal_email_or_log_failure,
-)
 from fractal_server.config import get_email_settings
 from fractal_server.config import get_settings
 from fractal_server.logger import close_logger
 from fractal_server.logger import set_logger
+from fractal_server.send_mail import send_fractal_email_or_log_failure
 from fractal_server.syringe import Inject
 
 logger = set_logger(__name__)
